requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
python-dotenv==1.1.0
requests==2.32.3
urllib3==2.4.0
pandas==2.2.2
scikit-learn==1.4.2
joblib==1.4.0
streamlit==1.34.0
